Un poco de refactor
