+++ projects/02-tic-tac-toe/src/components/Square.jsx
@@ -1,0 +1,13 @@
+export const Square = ({ children, isSelected, updateBoard, index }) => {
+  const classNameChecked = `square ${isSelected ? "is-selected" : ""}`;
+
+  const handleClick = () => {
+    updateBoard(index);
+  };
+
+  return (
+    <div className={classNameChecked} onClick={handleClick}>
+      {children}
+    </div>
+  );
+};
