+++ projects/02-tic-tac-toe/src/logic/board.js
@@ -1,0 +1,19 @@
+import { WINNER_COMBOS } from "../constants";
+
+export const checkWinnerFrom = (boardToCheck) => {
+  for (const combo of WINNER_COMBOS) {
+    const [a, b, c] = combo;
+    if (
+      boardToCheck[a] &&
+      boardToCheck[a] === boardToCheck[b] &&
+      boardToCheck[a] === boardToCheck[c]
+    ) {
+      return boardToCheck[a];
+    }
+  }
+  return null;
+};
+
+export const checkEndGame = (newBoard) => {
+  return newBoard.every((square) => square != null);
+};
